with long

+++ sort.java
@@ -1,7 +1,7 @@
 public class sort {
     public static void main(String[] args) {
         // This class is used to sort the files based on their content
-       short num=1000;
+        long num=1000L;
         short t = -200;
         System.out.println("Number of students :" +num) ;
         System.out.println("Number of teachers :" +t);
